backend/requirements/base.txt
# This file was autogenerated by uv via the following command:
#    ./bin/compile_dependencies.sh
amqp==5.2.0
    # via kombu
ape-pie==0.1.0
    # via zgw-consumers
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
asn1crypto==1.5.1
    # via webauthn
attrs==23.2.0
    # via
    #   glom
    #   jsonschema
    #   referencing
billiard==4.2.0
    # via celery
boltons==24.0.0
    # via
    #   face
    #   glom
cbor2==5.6.3
    # via webauthn
celery==5.4.0
    # via -r requirements/base.in
certifi==2024.7.4
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==43.0.1
    # via
    #   django-simple-certmanager
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
    #   webauthn
django==4.2.16
    # via
    #   -r requirements/base.in
    #   django-admin-index
    #   django-appconf
    #   django-axes
    #   django-cors-headers
    #   django-filter
    #   django-formtools
    #   django-hijack
    #   django-jsonform
    #   django-otp
    #   django-phonenumber-field
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-rosetta
    #   django-sendfile2
    #   django-simple-certmanager
    #   django-solo
    #   django-timeline-logger
    #   django-two-factor-auth
    #   djangorestframework
    #   drf-nested-routers
    #   drf-spectacular
    #   maykin-2fa
    #   mozilla-django-oidc
    #   mozilla-django-oidc-db
    #   zgw-consumers
django-admin-index==3.1.1
    # via -r requirements/base.in
django-appconf==1.0.6
    # via django-timeline-logger
django-axes==6.4.0
    # via -r requirements/base.in
django-cors-headers==4.3.1
    # via -r requirements/base.in
django-filter==24.2
    # via -r requirements/base.in
django-formtools==2.5.1
    # via django-two-factor-auth
django-hijack==3.4.5
    # via -r requirements/base.in
django-jsonform==2.23.0
    # via mozilla-django-oidc-db
django-ordered-model==3.7.4
    # via django-admin-index
django-otp==1.5.0
    # via django-two-factor-auth
django-phonenumber-field==7.3.0
    # via django-two-factor-auth
django-privates==2.0.0.post1
    # via
    #   -r requirements/base.in
    #   django-simple-certmanager
django-redis==5.4.0
    # via -r requirements/base.in
django-relativedelta==2.0.0
    # via zgw-consumers
django-rosetta==0.10.0
    # via -r requirements/base.in
django-sendfile2==0.7.1
    # via django-privates
django-simple-certmanager==2.0.0
    # via zgw-consumers
django-solo==2.2.0
    # via
    #   -r requirements/base.in
    #   mozilla-django-oidc-db
    #   zgw-consumers
django-timeline-logger==5.0.0
    # via -r requirements/base.in
django-two-factor-auth==1.16.0
    # via maykin-2fa
djangorestframework==3.15.2
    # via
    #   -r requirements/base.in
    #   djangorestframework-gis
    #   drf-nested-routers
    #   drf-spectacular
djangorestframework-camel-case==1.4.2
    # via -r requirements/base.in
djangorestframework-gis==1.0
    # via -r requirements/base.in
drf-nested-routers==0.94.1
    # via -r requirements/base.in
drf-spectacular==0.27.2
    # via -r requirements/base.in
ecs-logging==2.1.0
    # via elastic-apm
elastic-apm==6.22.0
    # via -r requirements/base.in
face==20.1.1
    # via glom
furl==2.1.3
    # via
    #   -r requirements/base.in
    #   ape-pie
glom==23.5.0
    # via mozilla-django-oidc-db
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-spectacular
josepy==1.14.0
    # via mozilla-django-oidc
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.7
    # via celery
maykin-2fa==1.0.0
    # via -r requirements/base.in
mozilla-django-oidc==4.0.1
    # via mozilla-django-oidc-db
mozilla-django-oidc-db==0.19.0
    # via -r requirements/base.in
orderedmultidict==1.0.1
    # via furl
phonenumberslite==8.13.35
    # via django-two-factor-auth
pillow==10.3.0
    # via -r requirements/base.in
polib==1.2.0
    # via django-rosetta
prompt-toolkit==3.0.43
    # via click-repl
psycopg2==2.9.9
    # via -r requirements/base.in
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via zgw-consumers
pyopenssl==24.2.1
    # via
    #   josepy
    #   webauthn
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   django-relativedelta
python-decouple==3.8
    # via -r requirements/base.in
python-dotenv==1.0.1
    # via -r requirements/base.in
python-slugify==8.0.4
    # via -r requirements/base.in
pyyaml==6.0.1
    # via drf-spectacular
qrcode==7.4.2
    # via django-two-factor-auth
redis==5.0.4
    # via django-redis
referencing==0.35.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   ape-pie
    #   django-rosetta
    #   mozilla-django-oidc
    #   zgw-consumers
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.11.0
    # via -r requirements/base.in
six==1.16.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.5.0
    # via django
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.11.0
    # via
    #   mozilla-django-oidc-db
    #   qrcode
    #   zgw-consumers
tzdata==2024.1
    # via celery
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.2
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.25.1
    # via -r requirements/base.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
webauthn==2.1.0
    # via django-two-factor-auth
wrapt==1.14.1
    # via elastic-apm
zgw-consumers==0.33.0
    # via -r requirements/base.in
